orders-api/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
pika
python-multipart
pyjwt
passlib[bcrypt]